Add vanilla css support to Separator (#517)
