+++ components/demos/index.jsx
@@ -16,7 +16,7 @@ export { default as ProgressDemo } from './Progress';
 export { default as RadioGroupDemo } from './RadioGroup';
 export { default as ScrollAreaDemo } from './ScrollArea';
 export { default as SelectDemo } from './Select';
-export { default as SeparatorDemo } from './Separator';
+export { default as SeparatorDemo } from './Separator/stitches';
 export { default as SliderDemo } from './Slider';
 export { default as SwitchDemo } from './Switch/stitches';
 export { default as TabsDemo } from './Tabs/stitches';
