+++ data/primitives/components/separator/1.0.1.mdx
@@ -13,9 +13,7 @@ aria: https://www.w3.org/TR/wai-aria-1.2/#separator
   <SeparatorDemo />
 </HeroContainer>
 
-```jsx hero template=Separator
-
-```
+<HeroCodeBlock folder="Separator" />
 
 <Highlights features={['Supports horizontal and vertical orientations.']} />
